🎨: 불필요한 코드 삭제 및 가독성을 위한 코드 구조 개선
