+++ iOS/todo-list/todo-list/UseCases/TaskManager.swift
@@ -73,33 +73,3 @@ struct TaskManager {
     }
 }
 
-class TaskStorage {
-    private var dict: [Int: Task] = [:]
-    
-    func all() -> [Task] {
-        return Array(dict.values)
-    }
-    
-    func overwrite(tasks: [Task]) {
-        dict = [:]
-        tasks.forEach { add(task: $0) }
-    }
-    
-    func add(task: Task) {
-        if dict[task.id] != nil {
-            return Logger.init().fault("\(task.id) already exists")
-        }
-        dict[task.id] = task
-    }
-    
-    func update(task: Task) {
-        if dict[task.id] == nil {
-            return Logger.init().fault("\(task.id) does not exists")
-        }
-        dict[task.id] = task
-    }
-    
-    func delete(id: Int) {
-        dict[id] = nil
-    }
-}
