+++ iOS/todo-list/todo-list/View/TaskTableViewCell.swift
@@ -10,7 +10,7 @@ import UIKit
 class TaskTableViewCell: UITableViewCell {
     static let identifier = "taskTableViewCell"
     
-    var id: Int?
+    var taskId: Int?
     
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
@@ -44,20 +44,17 @@ class TaskTableViewCell: UITableViewCell {
     }
     
     func configure(with viewModel: TaskCellViewModel) {
-        
-        self.id = viewModel.id
+        self.taskId = viewModel.id
         
         var config = defaultContentConfiguration()
+        var backgroundConfig = UIBackgroundConfiguration.listPlainCell()
         
         config.attributedText = self.makePrimaryText(title: viewModel.title)
-        
         config.secondaryAttributedText = self.makeSecondaryText(contents: viewModel.content, authored: "Authored by \(viewModel.device)")
         
-        self.contentConfiguration = config
-        
-        var backgroundConfig = UIBackgroundConfiguration.listPlainCell()
         backgroundConfig.cornerRadius = 10
         
+        self.contentConfiguration = config
         self.backgroundConfiguration = backgroundConfig
     }
     
@@ -66,7 +63,6 @@ class TaskTableViewCell: UITableViewCell {
     }
     
     func makeSecondaryText(contents: String, authored: String) -> NSAttributedString {
-        
         let contentText = NSMutableAttributedString.init(string: "\(contents)", attributes: Attributes.content)
         let newLine = NSAttributedString(string: "\n\n")
         let authoredText = NSMutableAttributedString.init(string: "\(authored)", attributes: Attributes.authored)
